notebook/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
requests>=2.31.0
matplotlib>=3.7.0 